more modifications for ssh script

+++ ssh.py
@@ -11,66 +11,77 @@ sys.path.append('./paramiko/lib-12.04/sites-packages/')
 import paramiko
 import ecdsa
 
-COMMAND='hostname; uptime'
 USER='secscan1'
+
+
+if len(sys.argv) == 2 and os.path.isfile(sys.argv[1]):
+  ssh_servers = open(sys.argv[1])
+else:
+  print "File needs to be entered with command. File needs to be valid."
+  exit(1)
+
 PASSWORD = getpass.getpass("Enter password for SSH systems:")
+PUBKEYIN = raw_input("Please enter your public key - if none - press enter")
+if (not os.path.isfile(PUBKEYIN)):
+  print "PUBLIC KEY not found."
+  exit(1)
+if not PUBKEYIN:
+  PUBKEY= "."
+PUBKEY = paramiko.RSAKey.from_private_key_file(PUBKEYIN)
 up = open("up_servers.txt",'w')
 
 def scan_ssh(HOST):
   try:
     client = paramiko.SSHClient()
     client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-    #print('*** Connecting to: %s ***' % HOST)
-    client.connect(hostname=HOST, port=22, username=USER, password=PASSWORD,timeout=5)
+    client.connect(hostname=HOST, port=22, username=USER, password=PASSWORD,\
+                   key_filename=PUBKEY,timeout=5)
     up.write("HOST: %s UP\n" % HOST)
-    #chan = client.invoke_shell()
-    #stdout = client.exec_command(COMMAND)
-    #chan.close()
     client.close()
 
   except gaierror as e:
-    #print('%s: Caught exception: %s: %s *** ' % (HOST,e.__class__, e))
     if (e[0] == 8):
       up.write("HOST: %s DNSINACTIVE\n" % HOST)
     else:
       up.write("HOST: %s :UNCAUGHT EXCEPT: %s\n" % (HOST,e))
+      up.write("\tException Class: %s\n" % e.__class__)
   
   except paramiko.AuthenticationException as e:
-    #print('%s: Caught exception: %s: %s *** ' % (HOST,e.__class__, e))
     up.write("HOST: %s AUTHFAIL\n" % HOST)
 
   except timeout as e:
-    #print('%s: Caught exception: %s: %s *** ' % (HOST,e.__class__, e))
     up.write("HOST: %s PORTBLOCKED\n" % HOST)
 
   except error as e:
-    #print('%s: Caught exception: %s: %s *** ' % (HOST,e.__class__, e))
     if (e[0] == 60):
       up.write("HOST: %s PORTBLOCKED\n" % HOST)
+    elif (e[0] == -2):
+      up.write("HOST: %s DNSLOOKUPFAIL\n" % HOST)
     else:
       up.write("HOST: %s :UNCAUGHT EXCEPT: %s\n" % (HOST,e))
+      up.write("\tException Class: %s\n" % e.__class__)
   
   except exceptions.NotImplementedError:
       up.write("HOST: %s URANDMISSING\n" % HOST)
-    
+
+  except paramiko.ssh_expcetion.SSHException as e:
+    up.write("HOST: %s FILTERING SSH\n" % HOST)
+
+  except exceptions.EOFError as e:
+    up.write("HOST: %s FILTERING SSH\n" % HOST)
 
   except Exception as e:
-    print('%s: Caught exception: %s: %s *** ' % (HOST,e.__class__, e))
-    up.write("HOST: %s :UNCAUGHT EXCEPT: %s\n" % (HOST,e))
-    try:
-      client.close()
-    except:
-      pass
+    up.write("HOST: %s :UNCAUGHT EXCEPTION: %s\n" % (HOST,e))
+    up.write("\tException Class: %s\n" % e.__class__)
+  try:
+    client.close()
+  except:
+    pass
   return
 
 
 def main():
   thread_list = []
-  if len(sys.argv) == 2 and os.path.isfile(sys.argv[1]):
-    ssh_servers = open(sys.argv[1])
-  else:
-    print "File needs to be entered with command. File needs to be valid."
-    exit(1)
   try:
     for line in ssh_servers:
       HOST = line.strip()
